feat: :sparkles: スタックの`id`と`stack_file_name`を`workspace.json`で持つように変更

+++ src-tauri/src/command/stack.rs
@@ -8,17 +8,25 @@ use crate::utils::CommandResult;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use thiserror::Error;
+use uuid::Uuid;
+
+use super::workspace::load_workspace;
+use super::workspace::update_workspace;
+use super::workspace::StackInfo;
+use super::workspace::Workspace;
 
 // フロントで利用する形
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Stack {
+    id: String,
     name: String,
     description_from_meta: Option<String>,
     description_from_stack: Option<String>,
+    exist: bool,
 }
 
 #[derive(Debug, Error)]
-enum LoadStackError {
+enum StackError {
     #[error("app error: {0}")]
     App(#[from] AppError),
     #[error("io error: {0}")]
@@ -27,9 +35,11 @@ enum LoadStackError {
     Yaml(#[from] serde_yaml::Error),
     #[error("json error: {0}")]
     Json(#[from] serde_json::Error),
+    #[error("workspace error: {0}")]
+    Workspace(#[from] super::workspace::WorkspaceError),
 }
 
-fn load_stack_meta(workspace_directory: &str, stack_name: &str) -> Result<Value, LoadStackError> {
+fn load_stack_meta(workspace_directory: &str, stack_name: &str) -> Result<Value, StackError> {
     // ${スタック名}.meta.jsonが存在するか確認
     let meta_path = format!("{}/{}.meta.json", workspace_directory, stack_name);
     let meta_path = Path::new(&meta_path);
@@ -37,7 +47,7 @@ fn load_stack_meta(workspace_directory: &str, stack_name: &str) -> Result<Value,
         // ${スタック名}.meta.jsonを作成する
         fs::File::create(&meta_path)?;
         // 空のJSONを作成
-        let empty_json = "{}".to_string();
+        let empty_json = format!("{{\"name\":\"{}\"}}", stack_name);
         fs::write(&meta_path, empty_json)?;
     }
 
@@ -53,31 +63,26 @@ fn load_stack_meta(workspace_directory: &str, stack_name: &str) -> Result<Value,
  *  2. ${スタック名}.meta.jsonを取得する（description_from_metaと各リソースとそのパラメータの説明文）
  *  3. メタファイルのdescriptionフィールドを取得する。Optionalな場合もある。（description_from_meta）
  */
-fn load_stacks(workspace_directory: &str) -> Result<Vec<Stack>, LoadStackError> {
+async fn load_stacks(workspace_directory: &str) -> Result<Vec<Stack>, StackError> {
+    // TODO: ディレクトリ配下のファイルをglobmatchで取得して、`stacks`に存在しないファイルも自動的に取り込む機能を追加する
     // globmatchを使って、workspace_directoryの直下にある*.template.jsonを取得する
-    let builder = globmatch::Builder::new("./*.template.json").build(workspace_directory);
-    let builder = match builder {
-        Ok(builder) => builder,
-        Err(_) => {
-            return Err(LoadStackError::App(AppError::new("Failed to load stacks")));
-        }
-    };
-    let paths: Vec<_> = builder.into_iter().flatten().collect();
+    // let builder = globmatch::Builder::new("./*.template.json").build(workspace_directory);
+    // let builder = match builder {
+    //     Ok(builder) => builder,
+    //     Err(_) => {
+    //         return Err(LoadStackError::App(AppError::new("Failed to load stacks")));
+    //     }
+    // };
+    // let paths: Vec<_> = builder.into_iter().flatten().collect();
+    let workspace = load_workspace(workspace_directory).await?;
     let mut stacks = Vec::new();
-    for path in paths {
-        // スタック名を取得する
-        let filename = path
-            .file_name()
-            .unwrap_or_default()
-            .to_str()
-            .unwrap_or_default()
-            .replace(".template.json", "");
-
+    for stack in workspace.stacks.iter() {
         // スタックを読み込む
-        let stack = match load_stack(workspace_directory, &filename) {
+        let stack = match load_stack_from_info(workspace_directory, stack).await {
             Ok(stack) => stack,
-            Err(_) => {
-                return Err(LoadStackError::App(AppError::new("Failed to load stack")));
+            Err(err) => {
+                print!("error: {:?}", err);
+                return Err(StackError::App(AppError::new("Failed to load stack")));
             }
         };
 
@@ -87,16 +92,41 @@ fn load_stacks(workspace_directory: &str) -> Result<Vec<Stack>, LoadStackError> 
     return Ok(stacks);
 }
 
-fn delete_stack(workspace_directory: &str, stack_name: &str) -> Result<(), LoadStackError> {
-    fs::remove_file(format!(
-        "{}/{}.template.json",
-        workspace_directory, stack_name
-    ))?;
-    fs::remove_file(format!("{}/{}.meta.json", workspace_directory, stack_name))?;
-    return Ok(());
+async fn delete_stack(workspace_directory: &str, stack_id: &str) -> Result<(), StackError> {
+    let workspace = load_workspace(workspace_directory).await?;
+    for stack in workspace.stacks.iter() {
+        if stack.id == stack_id {
+            // スタックファイルとメタデータファイルを削除する
+            fs::remove_file(format!("{}/{}", workspace_directory, stack.stack_file_name))?;
+            fs::remove_file(format!(
+                "{}/{}",
+                workspace_directory,
+                stack
+                    .stack_file_name
+                    .clone()
+                    .replace(".template.json", ".meta.json")
+            ))?;
+
+            // workspace.jsonのstacksから削除する
+            let stacks = workspace
+                .stacks
+                .iter()
+                .filter(|s| s.id != stack_id)
+                .cloned()
+                .collect();
+            let workspace = Workspace {
+                name: workspace.name,
+                description: workspace.description,
+                stacks,
+            };
+            update_workspace(workspace_directory, workspace).await?;
+            return Ok(());
+        }
+    }
+    return Err(StackError::App(AppError::new("Stack not found")));
 }
 
-fn import_stack(workspace_directory: &str, stack_file_path: &str) -> Result<(), LoadStackError> {
+async fn import_stack(workspace_directory: &str, stack_file_path: &str) -> Result<(), StackError> {
     let stack_file_path_buf = PathBuf::from(stack_file_path);
     let filename = stack_file_path_buf
         .file_name()
@@ -123,41 +153,84 @@ fn import_stack(workspace_directory: &str, stack_file_path: &str) -> Result<(), 
         fs::copy(stack_file_path, copy_file_path)?;
     }
 
+    // workspace.jsonのstacksに追加する
+    let workspace = load_workspace(workspace_directory).await?;
+    let mut stacks = workspace.stacks.clone();
+    stacks.push(StackInfo {
+        id: Uuid::new_v4().to_string(),
+        stack_file_name: filename.to_string(),
+    });
+    let workspace = Workspace {
+        name: workspace.name,
+        description: workspace.description,
+        stacks,
+    };
+    update_workspace(workspace_directory, workspace).await?;
+
     return Ok(());
 }
 
-fn load_stack(workspace_directory: &str, stack_name: &str) -> Result<Stack, LoadStackError> {
-    let template_path = format!("{}/{}.template.json", workspace_directory, stack_name);
+async fn load_stack_from_info(
+    workspace_directory: &str,
+    stack_info: &StackInfo,
+) -> Result<Stack, StackError> {
+    let template_path = format!("{}/{}", workspace_directory, stack_info.stack_file_name);
 
     // テンプレートファイルのdescriptionフィールドを取得する。Optionalな場合もある。（description_from_stack）
     let template_json = fs::read_to_string(&template_path)?;
     let template_json: Value = serde_json::from_str(&template_json).unwrap_or_default();
-    let description_from_stack = template_json["Description"]
+    let description_from_stack: Option<String> = template_json["Description"]
         .as_str()
         .and_then(|desc| Some(desc.to_string()));
 
     // メタファイルのdescriptionフィールドを取得する。Optionalな場合もある。（description_from_meta）
-    let description_from_meta = match load_stack_meta(workspace_directory, stack_name) {
-        Ok(meta_json) => meta_json["description"]
-            .as_str()
-            .and_then(|desc: &str| Some(desc.to_string())),
+    let meta_json = match load_stack_meta(
+        workspace_directory,
+        stack_info
+            .stack_file_name
+            .clone()
+            .replace(".template.json", "")
+            .as_str(),
+    ) {
+        Ok(meta_json) => meta_json,
         Err(_) => {
-            return Err(LoadStackError::App(AppError::new(
-                "Failed to load stack meta",
-            )));
+            return Err(StackError::App(AppError::new("Failed to load stack meta")));
         }
     };
+    println!("meta_json: {:?}", meta_json);
+    let stack_name = meta_json["name"]
+        .as_str()
+        .and_then(|name: &str| Some(name.to_string()))
+        .unwrap_or(stack_info.stack_file_name.clone());
+    let description_from_meta = meta_json["description"]
+        .as_str()
+        .and_then(|desc: &str| Some(desc.to_string()));
 
     // アプリ返却用のデータ構造作成
     return Ok(Stack {
-        name: stack_name.to_string(),
+        id: stack_info.id.clone(),
+        name: stack_name,
         description_from_meta: description_from_meta,
         description_from_stack: description_from_stack,
+        exist: true,
     });
 }
 
+async fn load_stack_from_id(
+    workspace_directory: &str,
+    stack_id: &str,
+) -> Result<Stack, StackError> {
+    let workspace = load_workspace(workspace_directory).await?;
+    for stack in workspace.stacks.iter() {
+        if stack.id == stack_id {
+            return load_stack_from_info(workspace_directory, stack).await;
+        }
+    }
+    return Err(StackError::App(AppError::new("Stack not found")));
+}
+
 #[tauri::command]
-pub fn load_stacks_command(
+pub async fn load_stacks_command(
     window: tauri::Window,
 ) -> Result<CommandResult<Vec<Stack>>, CommandResult> {
     let window_state = match get_window_state(window) {
@@ -166,16 +239,16 @@ pub fn load_stacks_command(
             return Err(CommandResult::failed("Window state not found"));
         }
     };
-    return match load_stacks(window_state.workspace_directory.as_str()) {
+    return match load_stacks(window_state.workspace_directory.as_str()).await {
         Ok(stacks) => Ok(CommandResult::success(stacks)),
         Err(e) => Err(CommandResult::failed(e.to_string().as_str())),
     };
 }
 
 #[tauri::command(rename_all = "snake_case")]
-pub fn delete_stack_command(
+pub async fn delete_stack_command(
     window: tauri::Window,
-    stack_name: &str,
+    stack_id: &str,
 ) -> Result<CommandResult<()>, CommandResult> {
     let window_state = match get_window_state(window) {
         Some(state) => state,
@@ -183,14 +256,14 @@ pub fn delete_stack_command(
             return Err(CommandResult::failed("Window state not found"));
         }
     };
-    return match delete_stack(window_state.workspace_directory.as_str(), stack_name) {
+    return match delete_stack(window_state.workspace_directory.as_str(), stack_id).await {
         Ok(_) => Ok(CommandResult::success(())),
         Err(e) => Err(CommandResult::failed(e.to_string().as_str())),
     };
 }
 
 #[tauri::command(rename_all = "snake_case")]
-pub fn import_stack_command(
+pub async fn import_stack_command(
     window: tauri::Window,
     stack_file_path: &str,
 ) -> Result<CommandResult<()>, CommandResult> {
@@ -200,16 +273,16 @@ pub fn import_stack_command(
             return Err(CommandResult::failed("Window state not found"));
         }
     };
-    return match import_stack(window_state.workspace_directory.as_str(), stack_file_path) {
+    return match import_stack(window_state.workspace_directory.as_str(), stack_file_path).await {
         Ok(_) => Ok(CommandResult::success(())),
         Err(e) => Err(CommandResult::failed(e.to_string().as_str())),
     };
 }
 
 #[tauri::command(rename_all = "snake_case")]
-pub fn load_stack_command(
+pub async fn load_stack_command(
     window: tauri::Window,
-    stack_name: &str,
+    stack_id: &str,
 ) -> Result<CommandResult<Stack>, CommandResult> {
     let window_state = match get_window_state(window) {
         Some(state) => state,
@@ -217,7 +290,7 @@ pub fn load_stack_command(
             return Err(CommandResult::failed("Window state not found"));
         }
     };
-    return match load_stack(window_state.workspace_directory.as_str(), stack_name) {
+    return match load_stack_from_id(window_state.workspace_directory.as_str(), stack_id).await {
         Ok(stack) => Ok(CommandResult::success(stack)),
         Err(e) => Err(CommandResult::failed(e.to_string().as_str())),
     };
